+++ src/main/java/net/minecraftforge/eventbus/internal/EventBusImpl.java
@@ -51,7 +51,7 @@ public record EventBusImpl<T extends Event>(
 
     @Override
     @SuppressWarnings({"unchecked", "rawtypes"}) // T extends Event, so this is safe.
-    public EventListener addListener(byte priority, Consumer<T> listener) {
+    public EventListener addListener(int priority, Consumer<T> listener) {
         return addListener(
                 priority == Priority.MONITOR
                         ? new EventListenerImpl.MonitoringListener(eventType, (Consumer<Event>) (Consumer) listener)
